hw1/requirements.txt
alembic
sqlalchemy
psycopg2-binary
python-dotenv